Increment version to 0.2.0 Since 0.1.0, a volatile TTL like cache has been implemented to improve file access speeds, javascript has been made compliant with ES6, and a mix task for archiving all unarchived files has been added.
The cache will likely need improvements to be made into a true or
close-to true LRU algorithm, as it uses random eviction on overflow
combined with TTL eviction.

Javascript has been refactored and linted to be compliant with airbnb
style guidelines. Various performance improvements were made, such as
the usage of `.join("")` to do string concatenation, reduced use of
closures, greater use of constant variables and more.

In order to allow for convenient archiving, a mix task archive.files has
been created. This detects all files which have not been uploaded to
archive.org, and uploads them. This will enable convenient archival by
the hostmaster of an Eientei instance.