Update version to 0.3.0 This release of Eientei focuses primarily on restructuring the API.
A file array may now be uploaded using the "files" key, and individual
file response will be given for each file.

A file response will be a JSON structured in one of two ways:
1. In the event of a success, the response will contain:
    * A "success" field which will be set to true.
    * A "url" field containing the full URL to the uploaded file.
    * A "name" field which will contain the original file name.
2. In the event of a failure the response will contain:
    * A "success" field which will be set to false.
    * A "reason" field which will indicate the reason for failure.
    * A "name" field which will contain the original file name.

When uploading using the "file" key and an individual file, the complete
response will be a JSON object containing a single "file" field which will
be a single file response.

When uploading using the "files" key and multiple files, the complete
response will be a JSON object containing a single "files" field which will
contain an array of file responses.

Due to Eientei being in a development state, no form of soft-deprecation will be
used. This is a breaking change and should be addressed immediatly by
all third party tool maintainers.